Download file with .ovpn extension instead of .conf
Based on recent changes in Foris (turris/foris/foris-openvpn-plugin!7) and from
feedback, which we gathered on support. From the GUI, it should download
a file from .ovpn extension, so Apple devices (iOS) can download it
directly without renaming it from .conf.

I was able to test this change with Network Manager on Ubuntu 20.04 and
importing works.